Revert version number, the POM needs modification;
To be able to deploy to Maven Central, the POM needs to be modified, so
the version number has been rolled back. As no one has been able to pull
the library, this shouldn't be a problem. I'll do some refactoring
before releasing the final 1.0-version as well, to prevent future
expansion problems.